Extract out shared composable for import directions
